Rename base.rb to be resource_mixin for clarity
