+++ lib/inspec/aws.rb
@@ -1,3 +1,3 @@
 require 'aws-sdk'
-require_relative 'aws/base'
+require_relative 'aws/resource_mixin'
 require_relative 'aws/connection'
