+++ lib/inspec/aws/resource_mixin.rb
